requirements.txt
# ─────────────────────────────────────────────────────────────────
# WEB FRAMEWORK
# fastapi[all]       — includes email validation, form support, Swagger UI extras
# uvicorn[standard]  — includes uvloop (fast async event loop) and httptools
# ─────────────────────────────────────────────────────────────────
fastapi[all]
uvicorn[standard]

# ─────────────────────────────────────────────────────────────────
# DATABASE
# sqlalchemy[asyncio] — ORM: define tables as Python classes; [asyncio] enables async queries
# alembic             — migration tool: tracks schema changes over time (like git for your DB)
# asyncpg             — async PostgreSQL driver (required for FastAPI's async endpoints)
# psycopg2-binary     — sync PostgreSQL driver (required by Alembic and Celery)
# ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]
alembic
asyncpg
psycopg2-binary
pgvector

# ─────────────────────────────────────────────────────────────────
# KAFKA
# aiokafka    — async Kafka producer and consumer; used by the pipeline consumer (FastAPI)
# kafka-python — sync Kafka producer; used by Celery tasks (sync context, no event loop)
# ─────────────────────────────────────────────────────────────────
aiokafka
kafka-python

# ─────────────────────────────────────────────────────────────────
# TASK QUEUE
# celery[redis] — task queue; [redis] includes the Redis broker transport
# redis         — Python Redis client (also used directly for WebSocket tickets)
# ─────────────────────────────────────────────────────────────────
celery[redis]
redis

# ─────────────────────────────────────────────────────────────────
# ML / NLP
# sentence-transformers — loads Sentence-BERT models locally; no API calls, no cost
#                         Downloads all-MiniLM-L6-v2 (~90MB) on first use.
#                         Pulls in PyTorch as a dependency (~2GB total install).
#                         First 'docker compose up --build' will be slow because of this.
# umap-learn            — dimensionality reduction before HDBSCAN clustering.
#                         Reduces 384-dim embeddings to 5-dim to avoid the curse
#                         of dimensionality. Centroids are still computed in 384-dim.
# hdbscan               — density-based clustering for sub-theme discovery.
#                         Discovers cluster count automatically (no fixed K needed).
# vaderSentiment        — rule-based sentiment analysis for Reddit comments.
#                         Designed for social media text. No API calls, microsecond latency.
# numpy                 — array math used by the sub-theme discovery task.
# ─────────────────────────────────────────────────────────────────
sentence-transformers
umap-learn
hdbscan
vaderSentiment
numpy

# ─────────────────────────────────────────────────────────────────
# EXTERNAL APIS
# google-genai — official Gemini API client (new SDK, replaces deprecated google-generativeai)
# twilio              — official Twilio client (SMS delivery)
# praw                — Python Reddit API Wrapper (Reddit ingestion)
# feedparser          — parses RSS and Atom XML feeds (BBC, Reuters, etc.)
# ─────────────────────────────────────────────────────────────────
google-genai
groq
langchain
langchain-cohere
twilio
praw
feedparser

# ─────────────────────────────────────────────────────────────────
# AUTHENTICATION
# PyJWT                    — HS256 JWT verification for NextAuth tokens
# ─────────────────────────────────────────────────────────────────
PyJWT

# ─────────────────────────────────────────────────────────────────
# CONFIGURATION
# python-dotenv    — loads .env file into os.environ at startup
# pydantic-settings — reads env vars into typed Pydantic models (Settings class)
# ─────────────────────────────────────────────────────────────────
python-dotenv
pydantic-settings

# ─────────────────────────────────────────────────────────────────
# HTTP CLIENTS
# httpx    — async-capable HTTP client (use this in async FastAPI endpoints)
# requests — sync HTTP client (kept for the existing stub /fetch endpoint)
# ─────────────────────────────────────────────────────────────────
httpx
requests

# ─────────────────────────────────────────────────────────────────
# TESTING
# pytest           — test runner
# pytest-asyncio   — allows writing async test functions with 'async def test_...'
# ─────────────────────────────────────────────────────────────────
pytest
pytest-asyncio

# ─────────────────────────────────────────────────────────────────
# DEV TOOLS
# ruff — extremely fast Python linter and formatter (replaces flake8 + black)
# ─────────────────────────────────────────────────────────────────
ruff
